refactor(balances): legacy
provide unified status rather than copy

https://tillhub.atlassian.net/browse/IOS-2235
