+++ lib/v1/balances/index.js
@@ -25,3 +25,14 @@ module.exports.get = {
     }
   }
 }
+
+module.exports.legacy = {
+  create: {
+    request: require('./legacy/create.request'),
+    response: {
+      $id: 'https://schemas.tillhub.com/v1/balances.legacy.post.response.schema.json',
+      $schema: 'http://json-schema.org/draft-07/schema#',
+      ...commonResponse({ resultItems: require('./legacy/response') })
+    }
+  }
+}
