+++ lib/v1/balances/legacy/create.request.js
@@ -1,5 +1,6 @@
 const base = require('./base')
-const statusCreate = require('./embedded/status/create.request')
+const statusCreate = require('../../statuses/create.request')
+const { oneOf } = require('../../../helpers/payload-or-null')
 
 module.exports = {
   $id: 'https://schemas.tillhub.com/v1/balances.legacy.create.request.schema.json',
@@ -10,9 +11,9 @@ module.exports = {
     'date'
   ],
   properties: {
-    ...base
-  },
-  status: {
-    ...statusCreate
+    ...base,
+    status: oneOf({
+      ...statusCreate
+    })
   }
 }
