+++ lib/v1/balances/legacy/embedded/cash_units/base.js
@@ -1,4 +1,4 @@
-const { oneOf } = require('../../../helpers/payload-or-null')
+const { oneOf } = require('../../../../../helpers/payload-or-null')
 
 module.exports = {
   description: 'Counted cash units.',
