+++ lib/v1/balances/legacy/embedded/expenses/base.js
@@ -1,4 +1,4 @@
-const { oneOf } = require('../../../helpers/payload-or-null')
+const { oneOf } = require('../../../../../helpers/payload-or-null')
 
 module.exports = {
   description: 'Expenses grouped by expense fa_account',
