+++ lib/v1/balances/legacy/embedded/metadata.js
@@ -1,5 +1,5 @@
 const { oneOf } = require('../../../../helpers/payload-or-null')
-const currencyAmount = require('../../../../common/currencyAmount')
+const currencyAmount = require('../../../../common/currency_amount')
 
 module.exports = {
   description: 'Extended balance stats',
