+++ lib/v1/balances/legacy/embedded/payments/base.js
@@ -1,4 +1,4 @@
-const { oneOf } = require('../../../helpers/payload-or-null')
+const { oneOf } = require('../../../../../helpers/payload-or-null')
 
 module.exports = {
   description: 'Payments grouped by payment_option',
