+++ lib/v1/balances/legacy/embedded/revenues/base.js
@@ -1,4 +1,4 @@
-const { oneOf } = require('../../../helpers/payload-or-null')
+const { oneOf } = require('../../../../../helpers/payload-or-null')
 
 module.exports = {
   description: 'Revenues grouped by revenue fa_account',
