--- lib/v1/balances/legacy/embedded/status/create.request.js
@@ -1,10 +1,0 @@
-const base = require('./base')
-
-module.exports = {
-  type: 'object',
-  additionalProperties: false,
-  required: [],
-  properties: {
-    ...base
-  }
-}
