+++ lib/v1/balances/legacy/embedded/tips/base.js
@@ -1,4 +1,4 @@
-const { oneOf } = require('../../../helpers/payload-or-null')
+const { oneOf } = require('../../../../../helpers/payload-or-null')
 
 module.exports = {
   description: 'Tips grouped by staff',
