+++ lib/v1/balances/legacy/embedded/vats/base.js
@@ -1,4 +1,4 @@
-const { oneOf } = require('../../../helpers/payload-or-null')
+const { oneOf } = require('../../../../../helpers/payload-or-null')
 const currencyAmountObject = require('../../../../../common/currency_amount')
 
 module.exports = {
