+++ lib/v1/balances/legacy/response.js
@@ -1,6 +1,6 @@
 const { oneOf } = require('../../../helpers/payload-or-null')
 const base = require('./base')
-const statusResponse = require('./embedded/status/response')
+const statusResponse = require('../../statuses/response')
 
 module.exports = {
   type: 'object',
